swap construction page for proper index
